requirements.txt
gradio==4.32.0
faiss-cpu==1.8.0.post1
dashscope==1.20.14
openai==1.55.3
httpx==0.27.0
llama-index-vector-stores-faiss==0.1.2
llama-index-embeddings-dashscope==0.1.4
llama-index-readers-file==0.1.33
docx2txt==0.8
openpyxl==3.1.5
llama-index-core==0.10.67
llama-index-llms-dashscope==0.1.2
llama-index-readers-dashscope==0.1.2
llama-index-llms-openai-like==0.1.3
uvicorn==0.30.6
fastapi==0.114.1
llama-index-postprocessor-dashscope-rerank-custom==0.1.0
simplejson==3.19.3
matplotlib==3.9.2
ragas==0.1.9
langchain_community==0.2.17
alibabacloud_bailian20231229==1.8.2
pandas==2.0.3
alibabacloud_green20220302==2.2.11
oss2==2.19.0
lagent==0.1.2
mmengine==0.10.3
ipywidgets==8.1.7
modelscope
flask==3.0.2
llama-index==0.10.20
openai==1.12.0
python-dotenv==1.0.0
dashscope==1.14.0
pandas==2.1.4
python-dateutil==2.8.2